+++ providers/youtube.php
@@ -17,6 +17,17 @@ class YouTube extends Provider {
 
 
   // ================================================
+  //  Parameters for Panel Field Cheatsheet
+  // ================================================
+
+  public function providerParameters() {
+    return [
+      ['t', 'Timecode where the video should start (e.g. 1m4s)'],
+    ];
+  }
+
+
+  // ================================================
   //  Autoplay
   // ================================================